added tablist ranks added stats api from pluginstube fixed that a empty message comes when quiting

+++ plugin/src/main/java/de/rexlmanu/mlgrush/plugin/player/GamePlayerData.java
@@ -17,7 +17,7 @@ import java.util.UUID;
 public class GamePlayerData implements Key {
 
   private UUID uniqueId;
-  private int coins;
+  private long coins;
   private List<String> boughtItems;
   private List<String> inventorySorting;
   private Statistics statistics;
